style:패키지 구조를 Layered Architecture 형식으로 변경
