+++ src/main/java/wooteco/subway/controller/StationController.java
@@ -1,8 +1,12 @@
-package wooteco.subway.station;
+package wooteco.subway.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.controller.dto.StationRequest;
+import wooteco.subway.controller.dto.StationResponse;
+import wooteco.subway.dao.StationDao;
+import wooteco.subway.domain.Station;
 import wooteco.subway.exception.DuplicateNameException;
 
 import java.net.URI;
