+++ src/main/java/wooteco/subway/dao/LineDao.java
@@ -1,4 +1,7 @@
-package wooteco.subway.line;
+package wooteco.subway.dao;
+
+import wooteco.subway.controller.dto.LineRequest;
+import wooteco.subway.domain.Line;
 
 import java.util.List;
 import java.util.Optional;
