+++ src/main/java/wooteco/subway/dao/LineJdbcDao.java
@@ -1,10 +1,12 @@
-package wooteco.subway.line;
+package wooteco.subway.dao;
 
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.controller.dto.LineRequest;
+import wooteco.subway.domain.Line;
 
 import java.sql.PreparedStatement;
 import java.util.List;
