+++ src/main/java/wooteco/subway/dao/StationFakeDao.java
@@ -1,6 +1,8 @@
-package wooteco.subway.station;
+package wooteco.subway.dao;
 
 import org.springframework.util.ReflectionUtils;
+import wooteco.subway.controller.dto.StationRequest;
+import wooteco.subway.domain.Station;
 import wooteco.subway.exception.DuplicateNameException;
 
 import java.lang.reflect.Field;
