+++ src/main/java/wooteco/subway/dao/StationJdbcDao.java
@@ -1,10 +1,12 @@
-package wooteco.subway.station;
+package wooteco.subway.dao;
 
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.controller.dto.StationRequest;
+import wooteco.subway.domain.Station;
 
 import java.sql.PreparedStatement;
 import java.util.List;
