+++ src/main/java/wooteco/subway/domain/Station.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.domain;
 
 public class Station {
     private final Long id;
