--- src/main/java/wooteco/subway/line/SectionRequest.java
@@ -1,25 +1,0 @@
-package wooteco.subway.line;
-
-public class SectionRequest {
-    private final Long upStationId;
-    private final Long downStationId;
-    private final int distance;
-
-    public SectionRequest(Long upStationId, Long downStationId, int distance) {
-        this.upStationId = upStationId;
-        this.downStationId = downStationId;
-        this.distance = distance;
-    }
-
-    public Long getUpStationId() {
-        return upStationId;
-    }
-
-    public Long getDownStationId() {
-        return downStationId;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-}
