+++ src/test/java/wooteco/subway/station/StationAcceptanceTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.controller.dto.StationResponse;
 
 import java.util.HashMap;
 import java.util.List;
@@ -20,153 +21,153 @@ import static org.assertj.core.api.Assertions.assertThat;
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
 
-	@DisplayName("지하철역을 생성한다.")
-	@Test
-	void createStation() {
-		// given
-		Map<String, String> params = new HashMap<>();
-		params.put("name", "강남역");
+    @DisplayName("지하철역을 생성한다.")
+    @Test
+    void createStation() {
+        // given
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "강남역");
 
-		// when
-		ExtractableResponse<Response> response = RestAssured.given()
-															.log()
-															.all()
-															.body(params)
-															.contentType(MediaType.APPLICATION_JSON_VALUE)
-															.when()
-															.post("/stations")
-															.then()
-															.log()
-															.all()
-															.extract();
+        // when
+        ExtractableResponse<Response> response = RestAssured.given()
+                                                            .log()
+                                                            .all()
+                                                            .body(params)
+                                                            .contentType(MediaType.APPLICATION_JSON_VALUE)
+                                                            .when()
+                                                            .post("/stations")
+                                                            .then()
+                                                            .log()
+                                                            .all()
+                                                            .extract();
 
-		// then
-		assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-		assertThat(response.header("Location")).isNotBlank();
-	}
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+    }
 
-	@DisplayName("기존에 존재하는 지하철역 이름으로 지하철역을 생성한다.")
-	@Test
-	void createStationWithDuplicateName() {
-		// given
-		Map<String, String> params = new HashMap<>();
-		params.put("name", "강남역");
-		RestAssured.given()
-				   .log()
-				   .all()
-				   .body(params)
-				   .contentType(MediaType.APPLICATION_JSON_VALUE)
-				   .when()
-				   .post("/stations")
-				   .then()
-				   .log()
-				   .all()
-				   .extract();
+    @DisplayName("기존에 존재하는 지하철역 이름으로 지하철역을 생성한다.")
+    @Test
+    void createStationWithDuplicateName() {
+        // given
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "강남역");
+        RestAssured.given()
+                   .log()
+                   .all()
+                   .body(params)
+                   .contentType(MediaType.APPLICATION_JSON_VALUE)
+                   .when()
+                   .post("/stations")
+                   .then()
+                   .log()
+                   .all()
+                   .extract();
 
-		// when
-		ExtractableResponse<Response> response = RestAssured.given()
-															.log()
-															.all()
-															.body(params)
-															.contentType(MediaType.APPLICATION_JSON_VALUE)
-															.when()
-															.post("/stations")
-															.then()
-															.log()
-															.all()
-															.extract();
+        // when
+        ExtractableResponse<Response> response = RestAssured.given()
+                                                            .log()
+                                                            .all()
+                                                            .body(params)
+                                                            .contentType(MediaType.APPLICATION_JSON_VALUE)
+                                                            .when()
+                                                            .post("/stations")
+                                                            .then()
+                                                            .log()
+                                                            .all()
+                                                            .extract();
 
-		// then
-		assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-	}
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+    }
 
-	@DisplayName("지하철역을 조회한다.")
-	@Test
-	void getStations() {
-		/// given
-		Map<String, String> params1 = new HashMap<>();
-		params1.put("name", "강남역");
-		ExtractableResponse<Response> createResponse1 = RestAssured.given()
-																   .log()
-																   .all()
-																   .body(params1)
-																   .contentType(MediaType.APPLICATION_JSON_VALUE)
-																   .when()
-																   .post("/stations")
-																   .then()
-																   .log()
-																   .all()
-																   .extract();
+    @DisplayName("지하철역을 조회한다.")
+    @Test
+    void getStations() {
+        /// given
+        Map<String, String> params1 = new HashMap<>();
+        params1.put("name", "강남역");
+        ExtractableResponse<Response> createResponse1 = RestAssured.given()
+                                                                   .log()
+                                                                   .all()
+                                                                   .body(params1)
+                                                                   .contentType(MediaType.APPLICATION_JSON_VALUE)
+                                                                   .when()
+                                                                   .post("/stations")
+                                                                   .then()
+                                                                   .log()
+                                                                   .all()
+                                                                   .extract();
 
-		Map<String, String> params2 = new HashMap<>();
-		params2.put("name", "역삼역");
-		ExtractableResponse<Response> createResponse2 = RestAssured.given()
-																   .log()
-																   .all()
-																   .body(params2)
-																   .contentType(MediaType.APPLICATION_JSON_VALUE)
-																   .when()
-																   .post("/stations")
-																   .then()
-																   .log()
-																   .all()
-																   .extract();
+        Map<String, String> params2 = new HashMap<>();
+        params2.put("name", "역삼역");
+        ExtractableResponse<Response> createResponse2 = RestAssured.given()
+                                                                   .log()
+                                                                   .all()
+                                                                   .body(params2)
+                                                                   .contentType(MediaType.APPLICATION_JSON_VALUE)
+                                                                   .when()
+                                                                   .post("/stations")
+                                                                   .then()
+                                                                   .log()
+                                                                   .all()
+                                                                   .extract();
 
-		// when
-		ExtractableResponse<Response> response = RestAssured.given()
-															.log()
-															.all()
-															.when()
-															.get("/stations")
-															.then()
-															.log()
-															.all()
-															.extract();
+        // when
+        ExtractableResponse<Response> response = RestAssured.given()
+                                                            .log()
+                                                            .all()
+                                                            .when()
+                                                            .get("/stations")
+                                                            .then()
+                                                            .log()
+                                                            .all()
+                                                            .extract();
 
-		// then
-		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-		List<Long> expectedLineIds = Stream.of(createResponse1, createResponse2)
-										   .map(it -> Long.parseLong(it.header("Location").split("/")[2]))
-										   .collect(Collectors.toList());
-		List<Long> resultLineIds = response.jsonPath()
-										   .getList(".", StationResponse.class)
-										   .stream()
-										   .map(StationResponse::getId)
-										   .collect(Collectors.toList());
-		assertThat(resultLineIds).containsAll(expectedLineIds);
-	}
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+        List<Long> expectedLineIds = Stream.of(createResponse1, createResponse2)
+                                           .map(it -> Long.parseLong(it.header("Location").split("/")[2]))
+                                           .collect(Collectors.toList());
+        List<Long> resultLineIds = response.jsonPath()
+                                           .getList(".", StationResponse.class)
+                                           .stream()
+                                           .map(StationResponse::getId)
+                                           .collect(Collectors.toList());
+        assertThat(resultLineIds).containsAll(expectedLineIds);
+    }
 
-	@DisplayName("지하철역을 제거한다.")
-	@Test
-	void deleteStation() {
-		// given
-		Map<String, String> params = new HashMap<>();
-		params.put("name", "강남역");
-		ExtractableResponse<Response> createResponse = RestAssured.given()
-																  .log()
-																  .all()
-																  .body(params)
-																  .contentType(MediaType.APPLICATION_JSON_VALUE)
-																  .when()
-																  .post("/stations")
-																  .then()
-																  .log()
-																  .all()
-																  .extract();
+    @DisplayName("지하철역을 제거한다.")
+    @Test
+    void deleteStation() {
+        // given
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "강남역");
+        ExtractableResponse<Response> createResponse = RestAssured.given()
+                                                                  .log()
+                                                                  .all()
+                                                                  .body(params)
+                                                                  .contentType(MediaType.APPLICATION_JSON_VALUE)
+                                                                  .when()
+                                                                  .post("/stations")
+                                                                  .then()
+                                                                  .log()
+                                                                  .all()
+                                                                  .extract();
 
-		// when
-		String uri = createResponse.header("Location");
-		ExtractableResponse<Response> response = RestAssured.given()
-															.log()
-															.all()
-															.when()
-															.delete(uri)
-															.then()
-															.log()
-															.all()
-															.extract();
+        // when
+        String uri = createResponse.header("Location");
+        ExtractableResponse<Response> response = RestAssured.given()
+                                                            .log()
+                                                            .all()
+                                                            .when()
+                                                            .delete(uri)
+                                                            .then()
+                                                            .log()
+                                                            .all()
+                                                            .extract();
 
-		// then
-		assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-	}
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
 }
